changed value from long to int

+++ src/common/kieker/common/util/registry/IIdRegistry.java
@@ -23,7 +23,7 @@ package kieker.common.util.registry;
  */
 public interface IIdRegistry<T> {
 
-	long get(T element);
+	int get(T element);
 
 	void putIfAbsent(T element);
 }
